chore: some rework
